bring sanity to the `render` interface
This commit does a number of things with the ultimate goal of
fixing the design of the `render` interface:

* most of the previous work of `render` moves to
  XML::Builder::Fragment::new

* `tag` and `tag_foreach` in XML::Builder go away; instead, it
  is now necessary to go through a namespace object

* the namespace object now instantiates qname objects, which
  can render to any of several string forms, but most
  importantly serve as the place for `tag` and `foreach`
  methods

* the `AUTOLOAD` stuff moves out of the NS class into
  a flyweight, nearly method-less Factory class

* XML::Builder loses the persisted prefix map

The ultimate point is that `render` now *requires* a namespace
to construct a tag. It also takes more than one argument and
will either construct a fragment or a tag, depending on
whether the first argument is a scalar reference (which scalar
is assumed to be the namespace).